+++ cmd/hangboard/internal/cmd/workout/testhang/flags.go
@@ -1,8 +1,9 @@
 package testhang
 
 import (
-	"github.com/spf13/cobra"
 	"time"
+
+	"github.com/spf13/cobra"
 )
 
 const (
@@ -10,6 +11,6 @@ const (
 )
 
 func flags(cmd *cobra.Command) error {
-	cmd.Flags().DurationP(flagDuration, "d", 12 * time.Second, "time interval for max hang test")
+	cmd.Flags().DurationP(flagDuration, "d", 12*time.Second, "time interval for max hang test")
 	return nil
 }
